+++ muutils/zanj/readme.md
@@ -13,6 +13,9 @@ a file `<filename>.zanj` is a zip file containing:
 	- when an element is too big, it can be moved to an external file
 		- `.npy` for numpy arrays or torch tensors
 		- `.jsonl` for pandas dataframes or large sequences
+	- list of external files stored in `__zanj_meta__.json`
+	- "$ref" key will have value pointing to external file
+	- `__format__` key will detail an external format type
 
 
 # Implementation
